Modified script

+++ test.sh
@@ -2,20 +2,29 @@
 
 # Set up the network settings based on the operating system
 # Install the mitmproxy certificate
-openssl x509 -in ~/.mitmproxy/mitmproxy-ca-cert.pem -out mitmproxy-ca-cert.crt
+if [[ -f "mitmproxy-ca-cert.crt" ]]; then
+    openssl x509 -in ~/.mitmproxy/mitmproxy-ca-cert.pem -out mitmproxy-ca-cert.crt
+fi
 
 # Install certificate and set up proxies for Mac
 if [[ "$OSTYPE" == "darwin"* ]]; then
+
     # Add in the mitmproxy certificate
-    sudo security add-trusted-cert -d -p ssl -p basic -k /Library/Keychains/System.keychain mitmproxy-ca-cert.crt
-    
+    if [[ -f "mitmproxy-ca-cert.crt" ]]; then
+        sudo security add-trusted-cert -d -p ssl -p basic -k /Library/Keychains/System.keychain mitmproxy-ca-cert.crt
+    fi
+
     networksetup -setwebproxy "Wi-Fi" localhost 8080
     networksetup -setsecurewebproxy "Wi-Fi" localhost 8080
+
 # Install certificate and set up proxies for Linux
 elif [[ "$OSTYPE" == "linux-gnu"* ]]; then
+
     # Add in mitmproxy certificate for local user
-    sudo cp mitmproxy-ca-cert.crt /usr/local/share/ca-certificates
-    sudo update-ca-certificates
+    if [[ -f "mitmproxy-ca-cert.crt" ]]; then
+        sudo cp mitmproxy-ca-cert.crt /usr/local/share/ca-certificates
+        sudo update-ca-certificates
+    fi
 
     export http_proxy=http://localhost:8080/
     export https_proxy=http://localhost:8080/
@@ -25,13 +34,14 @@ else
 fi
 
 # Code that runs the traffic scanners in the background
+sudo mitmproxy -s traffic_parser/mitmproxy_script.py &
 sudo python3 traffic_parser/dnsscan.py &
 # sudo python3 httpsscan.py &
-sudo mitmproxy -s traffic_parser/mitmproxy_script.py &
-python3 crawler/crawler.py crawler/sample_urls.txt 
-# python3 crawler
 
-End code cleanup: Remove the proxies and certificates
+# Run the crawler
+python3 crawler/crawler_dfs.py sample_urls.txt 0 9
+
+#End code cleanup: Remove the proxies and certificates
 if [[ "$OSTYPE" == "darwin"* ]]; then
     # sudo security delete-certificate -c "mitmproxy" /Library/Keychains/System.keychain
     networksetup -setwebproxystate "Wi-Fi" off
